First pass at making the LCP license functionality a library.
