+++ lcpserver/server/server.go
@@ -111,9 +111,9 @@ func New(bindAddr string, readonly bool, idx *index.Index, st *storage.Store, ls
 		// put content to the storage
 		s.handlePrivateFunc(contentRoutes, "/{content_id}", apilcp.AddContent, basicAuth).Methods("PUT")
 		// generate a license for given content
-		s.handlePrivateFunc(contentRoutes, "/{content_id}/license", apilcp.GenerateLicense, basicAuth).Methods("POST")
+		s.handlePrivateFunc(contentRoutes, "/{content_id}/license", apilcp.GenerateLicenseHandler, basicAuth).Methods("POST")
 		// deprecated, from a typo in the lcp server spec
-		s.handlePrivateFunc(contentRoutes, "/{content_id}/licenses", apilcp.GenerateLicense, basicAuth).Methods("POST")
+		s.handlePrivateFunc(contentRoutes, "/{content_id}/licenses", apilcp.GenerateLicenseHandler, basicAuth).Methods("POST")
 		// generate a licensed publication
 		s.handlePrivateFunc(contentRoutes, "/{content_id}/publication", apilcp.GenerateLicensedPublication, basicAuth).Methods("POST")
 		// deprecated, from a typo in the lcp server spec
@@ -127,8 +127,8 @@ func New(bindAddr string, readonly bool, idx *index.Index, st *storage.Store, ls
 
 	s.handlePrivateFunc(sr.R, licenseRoutesPathPrefix, apilcp.ListLicenses, basicAuth).Methods("GET")
 	// get a license
-	s.handlePrivateFunc(licenseRoutes, "/{license_id}", apilcp.GetLicense, basicAuth).Methods("GET")
-	s.handlePrivateFunc(licenseRoutes, "/{license_id}", apilcp.GetLicense, basicAuth).Methods("POST")
+	s.handlePrivateFunc(licenseRoutes, "/{license_id}", apilcp.GetLicenseHandler, basicAuth).Methods("GET")
+	s.handlePrivateFunc(licenseRoutes, "/{license_id}", apilcp.GetLicenseHandler, basicAuth).Methods("POST")
 	// get a licensed publication via a license id
 	s.handlePrivateFunc(licenseRoutes, "/{license_id}/publication", apilcp.GetLicensedPublication, basicAuth).Methods("POST")
 	if !readonly {
